Fix web tests due to redux-first-history migration
