+++ packages/web/package.json
@@ -184,7 +184,7 @@
     "react-virtualized": "9.22.3",
     "redux": "4.1.1",
     "redux-devtools-extension": "2.13.8",
-    "redux-first-history": "^5.2.0",
+    "redux-first-history": "5.2.0",
     "redux-persist": "6.0.0",
     "redux-saga": "1.1.3",
     "redux-sentry-middleware": "0.1.8",
