+++ packages/web/src/app/HistoryProvider.tsx
@@ -1,8 +1,15 @@
-import { createContext, memo, useContext } from 'react'
+import { createContext, memo, useContext, useMemo } from 'react'
 
-import { History } from 'history'
+import {
+  History,
+  BrowserHistoryBuildOptions,
+  createBrowserHistory,
+  createHashHistory,
+  HashHistoryBuildOptions,
+  createMemoryHistory
+} from 'history'
 
-import { createHistory } from 'utils/history'
+import { env } from 'services/env'
 
 export type HistoryContextType = {
   history: History
@@ -16,11 +23,29 @@ export const HistoryContext = createContext<HistoryContextType>({
   history: null as any
 })
 
-const history = createHistory()
+const USE_HASH_ROUTING = env.USE_HASH_ROUTING
+const basename = env.BASENAME
 
-// TODO: could put getPathname in here
 export const HistoryContextProvider = memo(
   (props: { children: JSX.Element }) => {
+    const history = useMemo(() => {
+      if (process.env.NODE_ENV === 'test') {
+        return createMemoryHistory()
+      } else if (USE_HASH_ROUTING) {
+        const config: HashHistoryBuildOptions = {}
+        if (basename) {
+          config.basename = basename
+        }
+        return createHashHistory(config)
+      } else {
+        const config: BrowserHistoryBuildOptions = {}
+        if (basename) {
+          config.basename = basename
+        }
+        return createBrowserHistory(config)
+      }
+    }, [])
+
     return (
       <HistoryContext.Provider value={{ history }}>
         {props.children}
